Added 2 folder for different codes.
OBSS is a company that made us a competition in Istanbul University and the code inside that folder will be the 2nd stage's code that i wrote.
Codility is a platform to try some algorithm questions. There is other things too but i will use it to find myself some algorithm questions.
I will be adding HackerRank if i solve a question that deserve downloading one day.